// apps/client/src/app/probable-waffle/game/entity/components/actor-audio/sound-type.ts
export enum SoundType {
  Select = "select",
  SelectExtra = "selectExtra",
  Move = "move",
  LocationUnavailable = "locationUnavailable",
  Attack = "attack",
  Build = "build",
  Chop = "chop",
  Mine = "mine",
  Repair = "repair",
  Heal = "heal",
  EnterContainer = "enterContainer",
  Damage = "damage",
  Death = "death",
  Stop = "stop"
}